Extract platform specific implementations into separate mods
